Simplify the cloud formation template
Change the "Subdomain" parameter to "Domain". This way, the full domain name does not have to be concatenated using the hosted zone and the subdomain. The domain and the hosted zone have partially redundant content. However, the production environment can now use the root domain name without having to specify a subdomain.